# Remote-write: reuse memory for marshalling