By holding a `proto.Buffer` per shard and passing it down to where marshalling is done, we avoid creating a lot of garbage.

Prometheus is using "gogo-proto", so I checked that the same API is available in regular: https://pkg.go.dev/github.com/golang/protobuf/proto#Buffer.Marshal

Basically all of the memory allocation in the benchmark is now on the decode side:
```
name              old time/op    new time/op    delta
SampleDelivery-4    59.7ms ± 3%    58.3ms ± 3%   -2.35%  (p=0.151 n=5+5)

name              old alloc/op   new alloc/op   delta
SampleDelivery-4    47.0MB ± 0%    38.8MB ± 0%  -17.38%  (p=0.008 n=5+5)

name              old allocs/op  new allocs/op  delta
SampleDelivery-4      412k ± 0%      412k ± 0%     ~     (p=0.817 n=5+5)
```

Also one drive-by improvement to a comment.